backend/requirements.txt

fastapi>=0.104.0
uvicorn[standard]>=0.23.2
asyncpg>=0.28.0
sqlalchemy>=2.0.0
pydantic-settings>=2.0.0
python-multipart>=0.0.6
aiofiles>=23.0.0
yt-dlp>=2023.7.6